requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
bitarray==3.4.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.1.1
cryptography==44.0.0
cytoolz==1.0.1
defusedxml==0.8.0rc2
Django==5.1.6
django-admin-soft-dashboard==1.0.26
django-cors-headers==4.7.0
django-environ==0.12.0
django-jazzmin==3.0.1
django-nested-admin==4.1.1
django-paypal==2.1
django-smart-selects==1.7.2
django-unfold==0.55.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
frozenlist==1.7.0
gunicorn==23.0.0
hexbytes==1.3.1
idna==3.10
multidict==6.5.0
mysqlclient==2.2.7
oauthlib==3.2.2
packaging==25.0
parsimonious==0.10.0
pillow==11.1.0
propcache==0.3.2
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyproject-toml==0.1.0
python-decouple==3.8
python-monkey-business==1.1.0
python3-openid==3.2.0
pytz==2025.1
pyunormalize==16.0.0
pywin32==310
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rlp==4.1.0
setuptools==79.0.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
toolz==1.0.0
types-requests==2.32.4.20250611
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.3.0
web3==7.12.0
websockets==15.0.1
wheel==0.45.1
whitenoise==6.9.0
yarl==1.20.1
